+++ tests/testthat/test-tar_render_manuscript.R
@@ -16,10 +16,17 @@ tar_test("tar_render_manuscript works for qmd input", {
   expect_equal(out, tibble::tibble(from = character(), to = character()))
   # results
   suppressWarnings(tar_make(callr_function = NULL))
-  expect_equal(fs::path_abs(tar_read(manuscript)[[1]]),
-               fs::path_real("output/manuscript.docx"))
-  expect_equal(fs::path_file(tar_read(manuscript)[[2]]),
+  expect_setequal(fs::path_abs(tar_read(manuscript)[1:2]),
+                  fs::path_real(c("output/manuscript.docx", "output/manuscript.pdf")))
+  expect_equal(fs::path_file(tar_read(manuscript)[[3]]),
                fs::path_file(qmd_file))
+  expect_setequal(
+    fs::path_abs(tar_read(manuscript)[4:6]),
+    fs::path_real(c("reports/references.bib", "reports/vancouver.csl",
+                    "reports/word-styles-reference-01.docx"))
+  )
+  expect_true(all(fs::file_exists(tar_read(manuscript)[1:2])))
+  expect_false(any(fs::file_exists(c("reports/manuscript.docx", "reports/manuscript.pdf"))))
   # Everything should be up to date.
   expect_equal(tar_outdated(callr_function = NULL), character(0))
 })